Extract login handler
