+++ frontend/packages/components/src/components/login/login.tsx
@@ -1,38 +1,19 @@
 'use client'
 
-import { LoginDocument, useMutation } from '@tekmetric/graphql'
 import { Button } from '@tekmetric/ui/button'
 import { Field } from '@tekmetric/ui/field'
 import { Form } from '@tekmetric/ui/form'
 import { Input } from '@tekmetric/ui/input'
 import { validateSchema } from '@tekmetric/ui/validate-schema'
 import { ValidationError } from '@tekmetric/ui/validation-error'
-import { useRouter } from 'next/navigation'
-import { useCallback } from 'react'
 
-import { useAuthStore } from '../../services/use-auth-store/use-auth-store'
 import { loginValidationSchema } from './constants'
+import { useLogin } from './services/use-login/use-login'
 import { LoginClassNames } from './styles'
 import type { LoginFormFields } from './types'
 
 export const Login = (): JSX.Element => {
-  const router = useRouter()
-  const [login, { error: globalError }] = useMutation(LoginDocument)
-  const { setSession } = useAuthStore((state) => state)
-  const hasGlobalError = Boolean(globalError)
-
-  const handleSubmit = useCallback(
-    async ({ email, password }: LoginFormFields) => {
-      const result = await login({ variables: { email, password } })
-
-      if (result.data?.login?.userId && !result.errors?.length) {
-        setSession({ userId: result.data.login.userId })
-
-        router.push(`/dashboard`)
-      }
-    },
-    [login, router, setSession]
-  )
+  const { handleSubmit, hasGlobalError } = useLogin()
 
   return (
     <Form<LoginFormFields>
